requirements.txt
anyio==3.6.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
fastapi==0.94.1
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
iniconfig==2.0.0
mysqlclient==2.1.1
packaging==23.0
pluggy==1.0.0
pydantic==1.10.6
pytest==7.2.2
python-dotenv==1.0.0
PyYAML==6.0
requests==2.28.2
rfc3986==1.5.0
sniffio==1.3.0
SQLAlchemy==2.0.6
SQLAlchemy-Utils==0.40.0
starlette==0.26.1
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.0
watchfiles==0.18.1
websockets==10.4
